fix: parsing support for json, console, and jsonselect
